adding exceptions, logger for errors and customizing http responses

+++ src/main/java/com/dh/apiDentalClinic/controller/AddressController.java
@@ -8,39 +8,53 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Collection;
-import java.util.Optional;
 
 @RequestMapping("/address")
 @RestController
 public class AddressController {
-@Autowired
+    @Autowired
     IAddressService addressService;
 
     @GetMapping("/all")
-    public Collection<AddressDTO> getAllAddress() {
-        return addressService.findAllAddress();
+    public ResponseEntity<Collection<AddressDTO>> getAllAddress() {
+
+        return ResponseEntity.ok(addressService.findAllAddress());
     }
 
     @GetMapping("/{id}")
-    public AddressDTO getAddress(@PathVariable Long id) {
-        return addressService.findAddressById(id);
+    public ResponseEntity<?> getAddress(@PathVariable Long id) {
+        AddressDTO addressDTO = addressService.findAddressById(id);
+        return ResponseEntity.ok(addressDTO);
     }
 
     @PostMapping("/add")
     public ResponseEntity<?> saveAddress(@RequestBody AddressDTO addressDTO) {
         addressService.saveAdrress(addressDTO);
-        return ResponseEntity.ok(HttpStatus.OK);
+        return ResponseEntity.ok("Address created successfully!!");
     }
 
     @PutMapping("/update")
     public ResponseEntity<?> updateAddress(@RequestBody AddressDTO addressDTO) {
-        addressService.updateAdrress(addressDTO);
-        return ResponseEntity.ok(HttpStatus.OK);
+        ResponseEntity<String> response;
+        if (addressService.findAddressById(addressDTO.getId()) != null) {
+            addressService.updateAdrress(addressDTO);
+            response = ResponseEntity.ok("Update address");
+        } else {
+            response = new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return response;
     }
 
     @DeleteMapping("/delete/{id}")
     public ResponseEntity<?> deleteAddress(@PathVariable Long id) {
-        addressService.deleteAdrress(id);
-        return ResponseEntity.ok(HttpStatus.OK);
+        ResponseEntity<String> response;
+        if (addressService.findAddressById(id) != null) {
+            addressService.deleteAdrress(id);
+            response = ResponseEntity.ok("Deleted address with id: " + id);
+        } else {
+            response = new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
+        return response;
     }
 }
